fix decimal number becoming integers on addition operation

+++ script.js
@@ -5,7 +5,7 @@ let operator = "";
 const res = document.querySelector("#result");
 
 function add(a, b){
-    return parseInt(a)+parseInt(b);
+    return parseFloat(a)+parseFloat(b);
 }
 
 function subtract(a, b){
